Moved the controller functions for handling Section Items out into their own controller class to help keep `ApplicationController` down to manageable size.
