+++ src/main/scala/rifs/business/controllers/ApplicationController.scala
@@ -99,98 +99,6 @@ class ApplicationController @Inject()(applications: ApplicationOps,
     }
   }
 
-
-  def deleteSectionItem(id: ApplicationId, sectionNumber: Int, itemNumber: Int) = Action.async { implicit request =>
-    applications.fetchAppWithSection(id, sectionNumber).flatMap {
-      case Some((app, os)) =>
-        val doc = os.map(_.answers).getOrElse(JsObject(Seq()))
-        val items = doc \ "items" match {
-          case JsDefined(JsArray(is)) => is.collect { case o: JsObject => o }
-          case _ => Seq()
-        }
-        val remainingItems = items.filterNot(o => hasItemNumber(o, itemNumber))
-        val updated = doc + ("items" -> JsArray(remainingItems))
-        applications.saveSection(id, sectionNumber, updated).map(_ => NoContent)
-
-      case None => Future.successful(NotFound)
-    }
-  }
-
-  def getSectionItem(id: ApplicationId, sectionNumber: Int, itemNumber: Int) = Action.async { implicit request =>
-    applications.fetchAppWithSection(id, sectionNumber).map {
-      case Some((app, os)) =>
-        val doc = os.map(_.answers).getOrElse(JsObject(Seq()))
-        jsonResult(findItem(doc, itemNumber))
-
-      case None => NotFound
-    }
-  }
-
-  private def hasItemNumber(o: JsObject, num: Int) = o \ "itemNumber" match {
-    case JsDefined(JsNumber(n)) if n == num => true
-    case _ => false
-  }
-
-  private def findItem(doc: JsObject, itemNumber: Int) = {
-    val items = doc \ "items" match {
-      case JsDefined(JsArray(is)) => is.collect { case o: JsObject => o }
-      case _ => Seq()
-    }
-    items.find(o => hasItemNumber(o, itemNumber))
-  }
-
-  def putSectionItem(id: ApplicationId, sectionNumber: Int, itemNumber: Int) = Action.async(parse.json[JsObject]) { implicit request =>
-    def hasItemNumber(o: JsObject, num: Int) = o \ "itemNumber" match {
-      case JsDefined(JsNumber(n)) if n == num => true
-      case _ => false
-    }
-
-    applications.fetchAppWithSection(id, sectionNumber).flatMap {
-      case Some((app, os)) =>
-        val doc = os.map(_.answers).getOrElse(JsObject(Seq()))
-        val items = doc \ "items" match {
-          case JsDefined(JsArray(is)) => is.collect { case o: JsObject => o }
-          case _ => Seq()
-        }
-        val remainingItems = items.filterNot(o => hasItemNumber(o, itemNumber))
-        val newItem = request.body + ("itemNumber" -> JsNumber(itemNumber))
-        val updated = doc + ("items" -> JsArray(newItem +: remainingItems))
-        applications.saveSection(id, sectionNumber, updated).map(_ => NoContent)
-
-      case None => Future.successful(NotFound)
-    }
-  }
-
-  def postSectionItem(id: ApplicationId, sectionNumber: Int) = Action.async(parse.json[JsObject]) { implicit request =>
-    applications.fetchAppWithSection(id, sectionNumber).flatMap {
-      case Some((app, os)) =>
-        val doc = os.map(_.answers).getOrElse(JsObject(Seq()))
-
-        val items: Seq[JsValue] = doc \ "items" match {
-          case JsDefined(JsArray(is)) => is
-          case _ => Seq()
-        }
-
-        allocateItemNumber().flatMap { n =>
-          val newItem = request.body + ("itemNumber" -> JsNumber(n))
-          val updated = doc + ("items" -> JsArray(newItem +: items))
-
-          applications.saveSection(id, sectionNumber, updated).map(_ => NoContent)
-        }
-
-      case None => Future.successful(NotFound)
-    }
-  }
-
-  /**
-    * TODO: use a db sequence for this
-    *
-    * @return
-    */
-  def allocateItemNumber(): Future[Int] = Future {
-    Random.nextInt().abs
-  }
-
   def completeSection(id: ApplicationId, sectionNumber: Int) = Action.async(parse.json[JsObject]) { implicit request =>
     applications.saveSection(id, sectionNumber, request.body, Some(DateTime.now(DateTimeZone.UTC))).map(_ => NoContent)
   }
